Add more web archiving docs
This commit changes some of the introductory material to cover some of
the things you should think about when choosing a tool for web
archiving. It also adds a new section for using browsertrix-crawler,
with an animated gif of the screencast feature, and also a screenshot of
ReplayWebPage for viewing WACZ archives.